requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3
decorator==4.4.2
idna==2.10
joblib==0.17.0
nltk==3.5
numpy==1.19.4
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.54.1
tzlocal==2.1
urllib3==1.26.2
